+++ algorithm/sort/source/SelectionSort.cpp
@@ -9,12 +9,12 @@
  * 
  *https://www.hackerearth.com/practice/algorithms/sorting/selection-sort/tutorial/
  */
-void SelectionSort(std::vector<uint64_t>& arr) {
-    uint64_t N = arr.size();
-    uint64_t min;
-    for (uint64_t i = 0; i < N-1; i++) {
+void SelectionSort(std::vector<size_t>& arr) {
+    size_t N = arr.size();
+    size_t min;
+    for (size_t i = 0; i < N-1; i++) {
         min = i;
-        for (uint64_t j = i+1; j < N; j++) {
+        for (size_t j = i+1; j < N; j++) {
             if (arr[j] < arr[min]) {
                 min = j;
             }
